Fix unnecessary or incorrect casts in vStringPut() calls
`vStringPut()` has been made forgiving in the type the character is
pass as in e6e018b90dcbcf6813243f700efc1ae5d4341ac5.  It however
requires the value to be valid in one of the formats it accepts.

This commits should update all problematic cases, basically removing
all explicit casts:

* removing incorrect `int` to `char` casts;
* removing incorrect `char` to `int` casts;
* removing unnecessary `char` to `unsigned char` casts.
* removing unnecessary `unsigned char` to `int` casts.

Incorrect casts can lead to unexpected results with values > 0x7F.
Unnecessary casts make the code harder to read than necessary, and
risk being incorrectly copied to code where they would be incorrect.

+++ parsers/abaqus.c
@@ -70,7 +70,7 @@ static void createTag(AbaqusKind kind, const char *buf)
 
 	do
 	{
-		vStringPut(name, (int) *buf);
+		vStringPut(name, *buf);
 		++buf;
 	} while ((*buf != '\0') && (*buf != ','));
 	makeSimpleTag(name, kind);
